Crate runqueue_priority

This crate contains the RunQueue structure, for priority scheduler. RunQueue structure is essentially a list of Tasks that it used for scheduling purposes.

Structs

PriorityTaskRef
A cloneable reference to a Taskref that exposes more methods related to task scheduling.

RunQueue
A list of references to Tasks (PriorityTaskRefs) that is used to store the Tasks (and associated scheduler related data) that are runnable on a given core. A queue is used for the token based prioirty scheduler. Runqueue implements Deref and DerefMut traits, which dereferences to VecDeque.

Constants

DEFAULT_PRIORITY

INITIAL_TOKENS

MAX_PRIORITY