Struct runqueue_priority::RunQueue

pub struct RunQueue { /* private fields */ }

A list of Tasks and their associated priority scheduler data that may be run on a given CPU core.

In rate monotonic scheduling, tasks are sorted in order of increasing periods. Thus, the period value acts as a form of task “priority”, with higher priority (shorter period) tasks coming first.

Implementations

impl RunQueue

pub fn init(which_core: u8, idle_task: TaskRef) -> Result<(), &'static str>

Creates a new RunQueue for the given core, which is an apic_id

pub fn get_runqueue( which_core: u8 ) -> Option<&'static PreemptionSafeRwLock<RunQueue>>

Returns RunQueue for the given core, which is an apic_id.

pub fn get_least_busy_core() -> Option<u8>

Returns the “least busy” core, which is currently very simple, based on runqueue size.

pub fn add_task_to_any_runqueue(task: TaskRef) -> Result<(), &'static str>

Chooses the “least busy” core’s runqueue (based on simple runqueue-size-based load balancing) and adds the given Task reference to that core’s runqueue.

pub fn add_task_to_specific_runqueue( which_core: u8, task: TaskRef ) -> Result<(), &'static str>

Convenience method that adds the given Task reference to given core’s runqueue.

pub fn remove_task(&mut self, task: &TaskRef) -> Result<(), &'static str>

Removes a TaskRef from this RunQueue.

pub fn remove_task_from_all(task: &TaskRef) -> Result<(), &'static str>

Removes a TaskRef from all RunQueues that exist on the entire system.

This is a brute force approach that iterates over all runqueues.

pub fn idle_task(&self) -> &TaskRef