Struct runqueue_priority::RunQueue

pub struct RunQueue { /* private fields */ }

A list of references to Tasks (PriorityTaskRefs) that is used to store the Tasks (and associated scheduler related data) that are runnable on a given core. A queue is used for the token based prioirty scheduler. Runqueue implements Deref and DerefMut traits, which dereferences to VecDeque.

Implementations

impl RunQueue

pub fn update_and_move_to_end(
&mut self,
index: usize,
tokens: usize
) -> Option<TaskRef>

Moves the TaskRef at the given index in this RunQueue to the end (back) of this RunQueue, and returns a cloned reference to that TaskRef. The number of tokens is reduced by one and number of context switches is increased by one. This function is used when the task is selected by the scheduler

pub fn init(which_core: u8) -> Result<(), &'static str>

Creates a new RunQueue for the given core, which is an apic_id

pub fn get_runqueue(which_core: u8) -> Option<&'static RwLockPreempt<RunQueue>>

Returns RunQueue for the given core, which is an apic_id.

pub fn get_least_busy_core() -> Option<u8>

Returns the “least busy” core, which is currently very simple, based on runqueue size.

pub fn add_task_to_any_runqueue(task: TaskRef) -> Result<(), &'static str>

Chooses the “least busy” core’s runqueue (based on simple runqueue-size-based load balancing) and adds the given Task reference to that core’s runqueue.

pub fn add_task_to_specific_runqueue(
which_core: u8,
task: TaskRef
) -> Result<(), &'static str>

Convenience method that adds the given Task reference to given core’s runqueue.

pub fn remove_task(&mut self, _task: &TaskRef) -> Result<(), &'static str>

Removes a TaskRef from this RunQueue.

pub fn remove_task_from_all(task: &TaskRef) -> Result<(), &'static str>

Removes a TaskRef from all RunQueues that exist on the entire system.

This is a brute force approach that iterates over all runqueues.

pub fn set_priority(task: &TaskRef, priority: u8) -> Result<(), &'static str>

Sets the priority of the given Task in all the RunQueue structures

pub fn get_priority(task: &TaskRef) -> Option<u8>

Output the priority of the given task. Outputs None if the task is not found in any of the runqueues.